feat: transfering front end project to parent root

+++ back-end/src/main/java/org/acme/dto/response/LabseqSequenceResponse.java
@@ -2,6 +2,6 @@ package org.acme.dto.response;
 
 
 
-public record LabseqSequenceResponse(int result, long executionTime) {
+public record LabseqSequenceResponse(long result, long executionTime) {
 
 }
